refactor: auth middleware (w.i.p.)
